AppDelegate의 vendingMachine 프로퍼티 관리로직을 별도 DataController 클래스로 이동
