+++ VendingMachineApp/VendingMachineApp/AppDelegate.swift
@@ -10,69 +10,33 @@ import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-    private var vendingMachine: VendingMachine! = VendingMachine.sharedInstance
-    private let encoder = JSONEncoder()
-    private let decoder = JSONDecoder()
+    private let dataController: VendingMachineDataController = VendingMachineDataController.shared
     var window: UIWindow?
     
-    private func encodeVendingMachineJsonData(){
-        encoder.outputFormatting = [.sortedKeys, .prettyPrinted]
-        guard let jsonData = try? encoder.encode(vendingMachine)else{
-            return
-        }
-        UserDefaults.standard.set(jsonData, forKey: "vendingMachine")
-    }
-    
-    private func flushPreviousVendingMachineData(){
-        UserDefaults.standard.set(nil, forKey: "vendingMachine")
-    }
-    
-    private func decodeJsonData(_ jsonData: Data)-> VendingMachine? {
-        guard let machine = try? decoder.decode(VendingMachine.self, from: jsonData) else{
-            return nil
-        }
-        return machine
-    }
-    
-    private func injectMachineToVendingViewController(){
+    /// injection
+    private func injectMachineToVendingViewController(_ vendingMachine: VendingMachine){
         guard let vendingVC = self.window?.rootViewController! as? VendingViewController else {
             return
         }
-        vendingVC.vendingMachine = vendingMachine
-    }
-    
-    private func decodeVendingMachineJsonData() {
-        guard let jsonData = loadData() else{
-            return
-        }
-        guard let machine = decodeJsonData(jsonData) else{
-            return
-        }
-        vendingMachine = machine
-        injectMachineToVendingViewController()
-        return
-    }
-    
-    private func loadData() -> Data? {
-        guard let jsonData = UserDefaults.standard.value(forKey: "vendingMachine") as? Data else{
-            return nil
-        }
-        return jsonData
+        vendingVC.configure(vendingMachine)
     }
     
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        decodeVendingMachineJsonData()
+        guard let vendingMachine = dataController.fetchData() else{ return true }
+        injectMachineToVendingViewController(vendingMachine)
         return true
     }
+    
     func applicationWillResignActive(_ application: UIApplication) {
     }
 
     func applicationDidEnterBackground(_ application: UIApplication) {
-        encodeVendingMachineJsonData()
+        dataController.encode()
     }
 
     func applicationWillEnterForeground(_ application: UIApplication) {
-        decodeVendingMachineJsonData()
+        guard let vendingMachine = dataController.fetchData() else{ return }
+        injectMachineToVendingViewController(vendingMachine)
     }
 
     func applicationDidBecomeActive(_ application: UIApplication) {
@@ -80,7 +44,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }
 
     func applicationWillTerminate(_ application: UIApplication) {
-        encodeVendingMachineJsonData()
+        dataController.encode()
     }
 }
